refactor(UserCreator, Dashboard): move logic associated with UserCreator Component to UserCreator module
