+++ src/components/UserCreator/UserCreator.consts.ts
@@ -6,6 +6,4 @@ export const defaultUserValues = {
 };
 
 export const NAME_OF_TEMP_APP = 'newApp';
-export const DASHBOARD = 'DASHBOARD';
-export const USER_CREATOR = 'USER_CREATOR';
 export const ERROR_ORIGIN = 'user creator';
