+++ src/components/UserCreator/index.ts
@@ -1,6 +1,1 @@
-export { default } from './UserCreator.view';
-export type {
-  IUserCreator,
-  IUserCreatorValues,
-  UserCreatorSubmitHandler,
-} from './UserCreator.types';
+export { default } from './UserCreator.container';
