--- src/page/Dashboard/Dashboard.types.ts
@@ -1,3 +1,0 @@
-import { IUserCreator } from 'components/UserCreator';
-
-export interface IDashboard { userCreatorProps: IUserCreator & { isLoading: boolean } }
