+++ src/page/Dashboard/Dashboard.view.tsx
@@ -1,27 +1,26 @@
-import Card from 'components/card';
+import AnswersManager from 'components/AnswersManager';
 import FormChooser from 'components/FormChooser';
+import PageContainer from 'components/PageContainer';
 import UserCreator from 'components/UserCreator';
 
 import useStyles from './Dashboard.styles';
-import { IDashboard } from './Dashboard.types';
 
 /**
  * Component contains multiple widget for user and form management.
  */
-export const Dashboard = ({ userCreatorProps }: IDashboard ): JSX.Element => {
+export const Dashboard = (): JSX.Element => {
   const styles = useStyles();
 
   return (
-    <div className={ styles.root }>
-      <div className={ styles.left }>
-        <FormChooser />
-        <Card
-          centerBody={ false }
-          isLoading={ userCreatorProps.isLoading }
-          body={ <UserCreator { ...userCreatorProps } /> }
-        />
+    <PageContainer>
+      <div className={ styles.root }>
+        <div className={ styles.left }>
+          <FormChooser />
+          <UserCreator />
+        </div>
+        <AnswersManager />
       </div>
-    </div>
+    </PageContainer>
   );
 };
